Fixed Bugs
Changes the 支出概览 子选项 to Chinese when in months, eg 一月 instead of january, which caused bugs of no data showing since the parser took in dates are parsed iinto system language which could be english. eg we filtered the expense by current year and selected month but chinese month doesn't match the month we have on hand parsed from data.json. we fixed that by adding for chinese and english checking and parsing